1Lab/services/app/requirements.txt
psycopg2-binary~=2.8.6 
memory-profiler~=0.58.0
